Minor changes to Hoarders, moved to Janitor cog
